Big Code Rewriting
Большая часть кода Переписана/Изменена

+++ app/src/main/java/tfre1t/example/pempogram/fragment/dashboard/addsound/Fragment_RecordSound.java
@@ -1,10 +1,10 @@
 package tfre1t.example.pempogram.fragment.dashboard.addsound;
 
+import android.content.Context;
 import android.graphics.PorterDuff;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.SystemClock;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -19,56 +19,58 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.constraintlayout.widget.Group;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
 
 import java.io.File;
 
-import tfre1t.example.pempogram.CheckPermission;
 import tfre1t.example.pempogram.R;
-import tfre1t.example.pempogram.database.DB;
 import tfre1t.example.pempogram.mediaplayer.MyMediaPlayer;
+import tfre1t.example.pempogram.trashсanclasses.CheckPermission;
+import tfre1t.example.pempogram.trashсanclasses.FillingCheck;
+import tfre1t.example.pempogram.ui.dashboard.DashboardViewModel;
 
 public class Fragment_RecordSound extends Fragment implements View.OnClickListener, SeekBar.OnSeekBarChangeListener, Chronometer.OnChronometerTickListener {
 
-    static final int START_RECORD = 1;
-    static final int STOP_RECORD = 2;
-    static final int PLAY_RECORD = 100;
-    static final int DELETE_CURRENT_RECORD = 200;
-    boolean isSave;
+    private static final int START_RECORD = 1;
+    private static final int STOP_RECORD = 2;
+    private static final int PLAY_RECORD = 100;
+    private static final int DELETE_CURRENT_RECORD = 200;
 
-    View v;
-    TextView tvTitle;
-    EditText etExecutorSound, etNameSound;
-    ImageButton imgBtnRecordAudiofile, imgBtnPlayAudiofile;
-    SeekBar sbAudiofile;
-    Group recordGroup, playGroup;
+    private DashboardViewModel dashboardViewModel;
+    private MyMediaPlayer myMediaPlayer;
+
+    private Context ctx;
+    private View v;
+    private File recordAudio;
+
+    private boolean isSave;
+    private int timeRecord;
+
+    private TextView tvTitle;
+    private EditText etExecutorSound, etNameSound;
+    private ImageButton imgBtnRecordAudiofile, imgBtnPlayAudiofile;
+    private SeekBar sbAudiofile;
+    private Group recordGroup, playGroup;
     private Chronometer tvTime,  tvRecordTime;
-
-    MyMediaPlayer myMediaPlayer;
-
-    DB db;
-    long id;
-    int timeRecord;
-
-    File recordAudio = null;
-
-    public Fragment_RecordSound(DB db, long i) {
-        this.db = db;
-        id = i;
-    }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        dashboardViewModel = new ViewModelProvider(getActivity()).get(DashboardViewModel.class);
         v = inflater.inflate(R.layout.fragment_addsound_recordsound, null);
-        onFindViewById();
+        ctx = v.getContext();
+
+        findViewById();
+
         isSave = false;
+        tvTitle.setText("Записать микрофон");
         return v;
     }
 
-    private void onFindViewById() {
+    private void findViewById() {
         tvTitle = v.findViewById(R.id.tvTitle);
-        tvTitle.setText("Записать микрофон");
         tvTime = v.findViewById(R.id.chTime);
+
         tvRecordTime = v.findViewById(R.id.chRecordTime);
         tvRecordTime.setOnChronometerTickListener(this);
 
@@ -80,8 +82,10 @@ public class Fragment_RecordSound extends Fragment implements View.OnClickListen
 
         imgBtnRecordAudiofile = v.findViewById(R.id.imgBtnRecordAudiofile);
         imgBtnRecordAudiofile.setOnClickListener(this);
+
         imgBtnPlayAudiofile = v.findViewById(R.id.imgBtnPlayAudiofile);
         imgBtnPlayAudiofile.setOnClickListener(this);
+
         v.findViewById(R.id.imgBtnDelRecord).setOnClickListener(this);
         v.findViewById(R.id.btnAdd).setOnClickListener(this);
 
@@ -91,55 +95,51 @@ public class Fragment_RecordSound extends Fragment implements View.OnClickListen
 
     @Override
     public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.imgBtnRecordAudiofile:
-                CheckPermission checkPermission = new CheckPermission();
-                //Проверка на разрешение использование микрофона
-                if (checkPermission.CheckPermissionRecord(getActivity()) || checkPermission.getCheckRecord()) {
-                    //myMediaPlayer == null записываем звук, если myMediaPlayer != null то останавливаем запись
-                    if (myMediaPlayer == null) {
-                        myMediaPlayer = new MyMediaPlayer();
-                        myMediaPlayer.startRecord(v.getContext());
-                        groupViewers(START_RECORD);
-                    } else {
-                        myMediaPlayer.stopRecord();
-                        groupViewers(STOP_RECORD);
-                        recordAudio = myMediaPlayer.getFileRecord();
-                    }
+        int id = v.getId();
+        if (id == R.id.imgBtnRecordAudiofile) {
+            CheckPermission checkPermission = new CheckPermission();
+            //Проверка на разрешение использование микрофона
+            if (checkPermission.CheckPermissionRecord(getActivity()) || checkPermission.getCheckRecord()) {
+                //myMediaPlayer == null записываем звук, если myMediaPlayer != null то останавливаем запись
+                if (myMediaPlayer == null) {
+                    myMediaPlayer = new MyMediaPlayer();
+                    myMediaPlayer.startRecord(ctx);
+                    groupViewers(START_RECORD);
+                } else {
+                    myMediaPlayer.stopRecord();
+                    groupViewers(STOP_RECORD);
+                    recordAudio = myMediaPlayer.getFileRecord();
                 }
-                break;
-            //Воспроизводим запись
-            case R.id.imgBtnPlayAudiofile:
-                myMediaPlayer.playRecord();
-                timeRecord = myMediaPlayer.getDuration();
-                groupViewers(PLAY_RECORD);
-                break;
-                //Удаляем запись
-            case R.id.imgBtnDelRecord:
-                myMediaPlayer.getFileRecord().delete();
-                myMediaPlayer = null;
-                groupViewers(DELETE_CURRENT_RECORD);
-                break;
-                //сохраняем запись
-            case R.id.btnAdd:
-                String nameSound = etNameSound.getText().toString();
-                String executorSound = etExecutorSound.getText().toString();
-                if (!fillingCheck(nameSound, executorSound, recordAudio)) {
-                    break;
-                }
-                String audiofile = recordAudio.getName();
-                db.addRecAudiofile(nameSound, executorSound, audiofile, id);
+            }
+        } //Воспроизводим запись
+        else if (id == R.id.imgBtnPlayAudiofile) {
+            myMediaPlayer.playRecord();
+            timeRecord = myMediaPlayer.getDuration();
+            groupViewers(PLAY_RECORD);
+        } //Удаляем запись
+        else if (id == R.id.imgBtnDelRecord) {
+            myMediaPlayer.getFileRecord().delete();
+            myMediaPlayer = null;
+            groupViewers(DELETE_CURRENT_RECORD);
+        } //сохраняем запись
+        else if (id == R.id.btnAdd) {
+            String nameSound = etNameSound.getText().toString();
+            String executorSound = etExecutorSound.getText().toString();
+            if (fillingCheck(nameSound, executorSound, recordAudio)) {
+                dashboardViewModel.addNewAudiofile(nameSound, executorSound, recordAudio.getName());
                 isSave = true;
-                Toast.makeText(v.getContext(), "Запись добавлена", Toast.LENGTH_SHORT).show();
-                requireActivity().setResult(1);
-                requireActivity().finish();
-                break;
+                Toast.makeText(ctx, "Запись добавлена", Toast.LENGTH_SHORT).show();
+                getActivity().finish();
+            }
         }
     }
 
     private void groupViewers(int status){
+        recordGroup.setVisibility(View.GONE);
+        playGroup.setVisibility(View.GONE);
         switch (status){
             case START_RECORD:
+                recordGroup.setVisibility(View.VISIBLE);
                 imgBtnRecordAudiofile.setImageResource(android.R.drawable.picture_frame);
                 imgBtnRecordAudiofile.getBackground().mutate().setColorFilter(getResources().getColor(android.R.color.holo_red_dark), PorterDuff.Mode.SRC_ATOP);
                 tvTime.setBase(SystemClock.elapsedRealtime());
@@ -149,11 +149,10 @@ public class Fragment_RecordSound extends Fragment implements View.OnClickListen
                 tvTime.stop();
                 tvRecordTime.setText(tvTime.getText());
                 tvTime.setText("00:00");
-
-                recordGroup.setVisibility(View.GONE);
                 playGroup.setVisibility(View.VISIBLE);
                 break;
             case PLAY_RECORD:
+                playGroup.setVisibility(View.VISIBLE);
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
                     tvRecordTime.setCountDown(true);
                     tvRecordTime.setBase(SystemClock.elapsedRealtime() + timeRecord);
@@ -161,34 +160,24 @@ public class Fragment_RecordSound extends Fragment implements View.OnClickListen
                 else {
                     tvRecordTime.setBase(SystemClock.elapsedRealtime() - timeRecord);
                 }
-                imgBtnPlayAudiofile.setImageResource(android.R.drawable.picture_frame);
+                imgBtnPlayAudiofile.setImageResource(R.drawable.baseline_stop_black_48);
                 sbAudiofile.setMax(timeRecord-1);
                 tvRecordTime.start();
                 break;
             case DELETE_CURRENT_RECORD:
                 tvRecordTime.stop();
-                imgBtnRecordAudiofile.setImageResource(android.R.drawable.ic_btn_speak_now);
+                imgBtnRecordAudiofile.setImageResource(R.drawable.baseline_mic_black_48);
                 imgBtnRecordAudiofile.getBackground().mutate().setColorFilter(getResources().getColor(R.color.colorSecondary), PorterDuff.Mode.SRC_ATOP);
                 recordGroup.setVisibility(View.VISIBLE);
-                playGroup.setVisibility(View.GONE);
                 break;
         }
     }
 
     private boolean fillingCheck(String nameSound, String executorSound, File audiofile) {
-        if (!nameSound.equals("")) {
-            etNameSound.getBackground().mutate().setColorFilter(getResources().getColor(R.color.colorTextPrimary), PorterDuff.Mode.SRC_ATOP);
-            if (!executorSound.equals("")) {
-                etExecutorSound.getBackground().mutate().setColorFilter(getResources().getColor(R.color.colorTextPrimary), PorterDuff.Mode.SRC_ATOP);
-                return audiofile != null;
-            } else {
-                etExecutorSound.getBackground().mutate().setColorFilter(getResources().getColor(android.R.color.holo_red_dark), PorterDuff.Mode.SRC_ATOP);
-                return false;
-            }
-        } else {
-            etNameSound.getBackground().mutate().setColorFilter(getResources().getColor(android.R.color.holo_red_dark), PorterDuff.Mode.SRC_ATOP);
-            return false;
-        }
+        FillingCheck fillCheck = new FillingCheck();
+        if (fillCheck.fillingCheckEditText(ctx, nameSound, etNameSound)) return false;
+        if (fillCheck.fillingCheckEditText(ctx, executorSound, etExecutorSound)) return false;
+        return audiofile != null;
     }
 
     @Override
@@ -197,7 +186,7 @@ public class Fragment_RecordSound extends Fragment implements View.OnClickListen
         if(!myMediaPlayer.mediaPlayerResume){
             tvRecordTime.stop();
             tvRecordTime.setText("00:00");
-            imgBtnPlayAudiofile.setImageResource(android.R.drawable.ic_media_play);
+            imgBtnPlayAudiofile.setImageResource(R.drawable.baseline_play_arrow_black_48);
         }
     }
 
@@ -219,7 +208,6 @@ public class Fragment_RecordSound extends Fragment implements View.OnClickListen
     @Override
     public void onDestroy() {
         if(myMediaPlayer != null){
-            Log.d("myLog", "onDestroy: isSave=" + isSave);
             if(!isSave){
                 myMediaPlayer.getFileRecord().delete();
             }
